+++ examples/pup/robotics/drivebase_basics.py
@@ -11,6 +11,9 @@ right_motor = Motor(Port.B)
 # The distance between the two wheel-ground contact points is 112mm.
 drive_base = DriveBase(left_motor, right_motor, wheel_diameter=56, axle_track=112)
 
+# Optionally, uncomment the line below to use the gyro for improved accuracy.
+# drive_base.use_gyro(True)
+
 # Drive forward by 500mm (half a meter).
 drive_base.straight(500)